+++ openpyxl2/cell/cell.py
@@ -106,6 +106,7 @@ class Cell(StyleableObject):
     def __init__(self, worksheet, column=None, row=None, value=None, col_idx=None, style_array=None):
         super(Cell, self).__init__(worksheet, style_array)
         self.row = row
+        """Row number of this cell (1-based)"""
         # _value is the stored value, while value is the displayed value
         self._value = None
         self._hyperlink = None
@@ -116,14 +117,17 @@ class Cell(StyleableObject):
         if column is not None:
             col_idx = column_index_from_string(column)
         self.col_idx = col_idx
+        """Column number of this cell (1-based)"""
 
 
     @property
     def coordinate(self):
+        """This cell's coordinate (ex. 'A5')"""
         return '%s%d' % (self.column, self.row)
 
     @property
     def column(self):
+        """The letter of this cell's column (ex. 'A')"""
         return get_column_letter(self.col_idx)
 
     @property
@@ -281,7 +285,7 @@ class Cell(StyleableObject):
     def value(self):
         """Get or set the value held in the cell.
 
-        :rtype: depends on the value (string, float, int or
+        :type: depends on the value (string, float, int or
             :class:`datetime.datetime`)
         """
         value = self._value
@@ -325,9 +329,9 @@ class Cell(StyleableObject):
 
     @property
     def is_date(self):
-        """Whether the value is formatted as a date
+        """True if the value is formatted as a date
 
-        :rtype: bool
+        :type: bool
         """
         return self.data_type == 'd' or (
             self.data_type == 'n' and is_date_format(self.number_format)
@@ -354,7 +358,7 @@ class Cell(StyleableObject):
     def comment(self):
         """ Returns the comment associated with this cell
 
-            :rtype: :class:`openpyxl2.comments.Comment`
+            :type: :class:`openpyxl.comments.Comment`
         """
         return self._comment